Rename regression_callables to regressors
